requirements.txt
flask
pymongo
bson
urllib
python-dotenv
datetime
onnxruntime
onnxmltools
onnx
onnxconverter-common
numpy
pandas
cohere
random
json
sqlalchemy
psycopg2-binary
gunicorn
sqlalchemy
psycopg2-binary
Pillow